Use separate file for gallery
