+++ src/sections/gallery/Gallery.module.css
@@ -1,0 +1,34 @@
+.slider {
+  aspect-ratio: 1 / 1;
+}
+
+.container {
+  position: relative;
+  overflow: hidden;
+  max-width: 100%;
+  z-index: 1;
+}
+.container::after {
+  content: "";
+  position: absolute;
+  left: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  background: #e5e5e5;
+  opacity: 0;
+  transition: 0.5s ease all;
+}
+img,
+iframe {
+  transition: 0.5s ease all;
+  width: 100%;
+  object-fit: cover;
+}
+.container:hover img {
+  transform: scale(1.1);
+}
+
+.container:hover::after {
+  opacity: 0.2;
+}
